requirements.txt
dj-database-url>=0.5.0
Django
gunicorn>=20.0.0,<21.0.0
psycopg2-binary>=2.8.0,<3.0.0
whitenoise>=5.2.0,<6.0.0
django-rest-framework